Renames o.e.s.event.queues.processors
Renames o.e.s.event.queues.processors to o.e.s.event.queues.processes. 
Updates all dependencies.
