+++ org.eclipse.scanning.event.queues/META-INF/MANIFEST.MF
@@ -12,6 +12,6 @@ Import-Package: org.osgi.framework;version="1.7.0",
  org.osgi.service.component;version="1.2.0",
  org.slf4j;version="[1.7.2,2.0.0)"
 Export-Package: org.eclipse.scanning.event.queues,
- org.eclipse.scanning.event.queues.processors,
+ org.eclipse.scanning.event.queues.processes,
  org.eclipse.scanning.event.queues.remote
 Bundle-ActivationPolicy: lazy
