+++ org.eclipse.scanning.event.queues/src/org/eclipse/scanning/event/queues/processes/TaskBeanProcess.java
@@ -1,4 +1,4 @@
-package org.eclipse.scanning.event.queues.processors;
+package org.eclipse.scanning.event.queues.processes;
 
 import org.eclipse.scanning.api.event.EventException;
 import org.eclipse.scanning.api.event.alive.PauseBean;
