+++ org.eclipse.scanning.test/src/org/eclipse/scanning/test/event/queues/processes/SubTaskAtomProcessorIntegrationTest.java
@@ -1,4 +1,4 @@
-package org.eclipse.scanning.test.event.queues.processors;
+package org.eclipse.scanning.test.event.queues.processes;
 
 import org.eclipse.scanning.api.event.IEventService;
 import org.eclipse.scanning.api.event.queues.IQueueControllerService;
